Transform-EaDiagramToVisio.ps1 - fixed shape deletion by not just removing shapes, if their corresponding EA element got deleted but also removing shapes, if their corresponding EA element is not part of the diagram anymore
fixes #21

+++ src/Transform-EaDiagramToVisio.ps1
@@ -249,7 +249,13 @@ PROCESS
 		$element = $eaElementsByElementGUID[$s.Data1];
 		if ($null -eq $element)
 		{
+			$s | Remove-Shape;
+			continue;
 		}
+		
+		# remove shape, if not exists in EA diagram
+		$diagramObject = $diagram.DiagramObjects |? ElementID -eq $s.ElementID;
+		if ($null -eq $diagramObject)
 		{
 			$s | Remove-Shape;
 		}
